Removed the entity interfaces, as they were too much cruft.
Documentation will enforce the immutability of added objects. It's weak but all those interfaces got annoying fast - not really worth the duplication of code.

At some point, this could be enforced by a "freezable" implementation for the elements but only if this implementation can be automated (which today it cannot).