Content key data is now optional when loading documents.
Key data is optional in CPIX v2.2 and in some use-cases (e.g. when using CPIX for key requests), the optionality is desirable.

However, the requirement was kept intact for the adding of new keys. This is to keep the library in line with the original/general CPIX vision (CPIX is for transporting content keys -- in which case not having key data would be suspect) and to avoid upsetting some of the existing functionality.