Add 256-bit key support
The library will now consider keys valid if they are either 128-bit or 256-bit in size. This allows the library to be used in scenarios where 256 bit keys are used.

The CPIX specification does not place restrictions on key size, so the 128-bit limit was was only a library-specific limitation.